requirements.txt
asgiref==3.6.0
Django==4.2
django-cors-headers==3.14.0
django-filter==23.1
django-rest-multiple-models==2.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
importlib-metadata==6.6.0
Markdown==3.4.3
Pillow==9.5.0
pkg_resources==0.0.0
psycopg==3.1.8
psycopg-binary==3.1.8
psycopg2==2.9.6
psycopg2-binary==2.9.6
PyJWT==1.7.1
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.5.0
zipp==3.15.0